# backend/docs/swagger.yaml
basePath: /
definitions:
...
  http.createTournamentRequest:
    properties:
      name:
        example: Spring Championship 2026
        type: string
      organizerId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      settings:
        $ref: '#/definitions/models.TournamentSettings'
    type: object
...
  http.deleteUserResponse:
    properties:
      message:
        example: user deleted
        type: string
    type: object
  http.errorResponse:
    properties:
      error:
        properties:
          code:
            example: INVALID_INPUT
            type: string
          message:
            example: validation failed
            type: string
        type: object
    type: object
...
  http.listUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/inbound.UserDTO'
        type: array
    type: object
...
  http.reportMatchRequest:
    properties:
      matchId:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      reportedBy:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      results:
        items:
          $ref: '#/definitions/inbound.MatchResult'
        type: array
    type: object
  http.updateTournamentRequest:
    properties:
      settings:
        $ref: '#/definitions/models.TournamentSettings'
    type: object
  http.updateUserRequest:
    properties:
      role:
        example: admin
        type: string
    type: object
  http.updateUserResponse:
    properties:
      message:
        example: user updated
        type: string
    type: object
  inbound.MatchResult:
    properties:
      placement:
        description: 1, 2, 3, or 4
        type: integer
      playerId:
        type: string
      seatColor:
        $ref: '#/definitions/models.SeatColor'
    type: object
  inbound.UserDTO:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  models.AdvancementConfig:
    properties:
      advancement_per_game:
        items:
          $ref: '#/definitions/models.AdvancementPerGame'
        type: array
      games:
        type: integer
      round:
        type: string
    type: object
  models.AdvancementPerGame:
    properties:
      game_ids:
        items:
          type: integer
        type: array
      placements:
        description: e.g., [1, 2] means 1st and 2nd advance
        items:
          type: integer
        type: array
    type: object
...
  models.Match:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      league_id:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      placement_points:
        items:
          type: integer
        type: array
      round:
        type: integer
      status:
        $ref: '#/definitions/models.MatchStatus'
      table_number:
        type: integer
      tournament_id:
        type: string
    type: object
  models.MatchStatus:
    enum:
    - pending
    - completed
    type: string
    x-enum-varnames:
    - MatchStatusPending
    - MatchStatusCompleted
...
  models.SeatColor:
    enum:
    - yellow
    - green
    - blue
    - red
    type: string
    x-enum-varnames:
    - SeatYellow
    - SeatGreen
    - SeatBlue
    - SeatRed
  models.Tournament:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      name:
        type: string
      organizer_id:
        type: string
      settings:
        $ref: '#/definitions/models.TournamentSettings'
      status:
        $ref: '#/definitions/models.TournamentStatus'
      type:
        type: string
    type: object
  models.TournamentSettings:
    properties:
      advancement:
        items:
          $ref: '#/definitions/models.AdvancementConfig'
        type: array
      default_reporter:
        description: '"lowest_advancing"'
        type: string
      tables_count:
        type: integer
    type: object
  models.TournamentStatus:
    enum:
    - draft
    - live
    - completed
    type: string
    x-enum-varnames:
    - TournamentStatusDraft
    - TournamentStatusLive
    - TournamentStatusComplete
host: localhost:8080
info:
  contact: {}
  description: API for managing Ludo tournaments and leagues
  title: Ludo Tournament Management API
  version: "1.0"
paths:
...
  /tournaments:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all tournaments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tournament'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: List all tournaments
      tags:
      - tournaments
    post:
      consumes:
      - application/json
      description: Creates a new knockout tournament with the given settings
      parameters:
      - description: Tournament details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createTournamentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tournament'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a new tournament
      tags:
      - tournaments
  /tournaments/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a tournament by its unique identifier
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a tournament
      tags:
      - tournaments
    get:
      consumes:
      - application/json
      description: Retrieves a tournament by its unique identifier
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tournament'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a tournament by ID
      tags:
      - tournaments
    patch:
      consumes:
      - application/json
      description: Updates an existing tournament's settings
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Tournament settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.updateTournamentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a tournament
      tags:
      - tournaments
  /tournaments/{id}/matches:
    get:
      consumes:
      - application/json
      description: Retrieves all matches for a specific tournament
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Match'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: List matches for a tournament
      tags:
      - tournaments
  /tournaments/{id}/pairings:
    get:
      consumes:
      - application/json
      description: Retrieves the pairings for the current round of a tournament
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current round pairings
      tags:
      - tournaments
  /tournaments/{id}/report:
    post:
      consumes:
      - application/json
      description: Reports the result of a match in a tournament
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Match result details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.reportMatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Report a match result
      tags:
      - tournaments
  /users:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all users (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.listUsersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a user by their unique identifier (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.deleteUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Updates an existing user's role (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.updateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
swagger: "2.0"